refactor(GameResult): GameResult 패키지 변경 및 Dealer 테스트 코드 변경
